Use niv to pin Nixpkgs, track nixos-unstable in system config
This is a (relatively) huge changeset, mostly because it involves
a fundamental change in how Nixpkgs and other dependencies in the
future are managed in this repo. References to unstable in the home
config have been removed, since we now track the nixos-unstable branch
of nixpkgs, and several things in the system config have also been
moved around for organization.

Tracking nixos-unstable should (theoretically) be safe, since we can
simply roll back any versions by reverting the version of nixpkgs
we pin. This makes for more of a "rolling release" system, but
with the benefits of determinism/rollbacks.

Also, the packageOverrides in nixpkgs.config have been replaced with
overlays that are imported in both the system- and home-level. An
"unstable-fixes" overlay was also added, so if anything breaks, I can
just grab the derivation that was last working. This is currently being
done for 'ldc', a compiler for the D language (and 'dmd', so I can
pull that from the binary cache without having to build it).

One potential problem: When nixpkgs gets updated, evaluating the config
involves updating the nix search path (more specifically /etc/sources),
which means the system currently evaluating is not using the same
nix path as the system being evaluated. To work around this, we can just
do another run of nixos-rebuild, but this doesn't seem ideal.
I'm also not entirely sure whether this poses much of an issue.