fish: Add z plugin, use bobthefish from sources
home-manager has merged plugin and function definitions for fish
upstream. Unfortunately, the way it sources fish files doesn't allow for
prompt plugins to be added properly, so I still have to source those
myself. Still, I've removed it from my overlay (doesn't make sense for
it to be there) and instead use the source directly in the fish config.