Use overlays in Nix path
This lets me use the overlays in Nix tools outside this config.

Since overlays/default.nix was just a list of overlays we used in the
nixpkgs.overlays option, we can get the same thing by just using map to
import each file in overlays/. So we also remove overlays/default.nix,
since it's not an actual overlay.