Add preliminary flake support
Flakes are essentially repos that define inputs (dependencies) and
outputs, which are in this case NixOS configurations for each of my
machines. This is a proposed RFC, but support for flakes has been merged
to nixpkgs and Nix 2.4. This commit turns this repo into a flake.

I hope to eventually replace tracking dependencies through niv with this,
since this is essentially built into Nix itself. For modules, this just
involves passing the flake's inputs to each module; for now, these inputs
are passed in but ignored. Overlays are a bit more tricky, since those
that depend on inputs will have to be defined in a separate module.

Having nixpkgs defined as a flake input rather than a niv dependency
(which was eventually symlinked to /etc/sources) will hopefully
solve the problem introduced in 80ec458, where I essentially have to
rebuild twice on a nixpkgs update.