Restructure flake, make configuration more modular
The entire structure of this flake has been redone, with much (like, a
LOT) of inspiration from hlissner's dotfiles flake. Most of the modules
in ./modules define options as well as config so that they can be
switched on and off per host.

The "device" argument has been removed, as it's no longer needed.
Modules in ./modules are enabled per-host in ./hosts, with baseline
profiles being defined in ./hosts/profiles.

The flake's output set has been extended to include:
- A library of functions I use throughout the config, which extends
  nixpkgs' library,
- The packages defined in ./packages, as well as an overlay containing
  them,
- The modules themselves, since they are pretty much self-contained and
  can be used outside of this flake
- The custom home-manager modules I defined and used in this config,
  which anyone else is free to import and use