Move profiles & explicit options into roles/
Roles are now used to define the high-level properties of a machine,
such as whether it is a server or a work machine. For now, I just have a
base role (that all machines will use) and a personal role. This is to
prepare for adding my minidepot server in and repurposing the minidepot
repo as a NixOS module to be imported instead.

Also: previously, some files in modules/ would explicitly set options,
whether they were enabled or not. All such options have been moved into
the appropriate role. Ideally, this means that if a configuration does
not import any roles, no options will be implicitly set.