nvim: Move extra config to top of init.vim
Adding extra configuration to the bottom (the default behavior in the
home-manager module) results in some plugin mappings being set before
base mappings are set. For example, the leader key is set in
mappings.vim, but this would be set after all plugins are configured, so
effectively all the <Leader> plugin mappings were mapped to '\'.

This fixes that, so now (among other things) <Leader> mappings are
appropriately mapped to <Space>.