bin: Add funcs for "dicts", add --yes to ask_yesno
Dictionaries are implemented as universal fish variables (for lack of a
better way to persist variables outside of a script). Implement methods
to set, get, query, and list dictionary keys.

Also adds a --yes/-y option to ask_yesno to change the default choice to
'y' instead of 'n', and make responses case-insensitive.