comma: Add caching for cmds with multiple choices
If I want to run the same command several times, and that command is
available in several packages, I don't want to have to select the same
package each time. This implements a cache in the form of universal
variables for such commands. comma will prompt to use a previously
selected choice if one was made before.