Add overlay for failing nixpkgs builds
Uses the nix-lastgoodrev script to get the last successful build from
Hydra and puts it in an overlay.

There is also a super extremely ugly hack involving
import-from-derivation and an unsafe function that checks whether a
package still needs to be in that overlay. I will probably end up
moving this functionality into the sysdo script instead.