treefmt: Use filter to find all fish scripts in bin
Uses `builtins.readFile` to check if the script is interpreted with
fish, which was the best thing I could come up with for such a filter.
This is better than having to manually update a list each time I add a
new script to /bin, but it does take slightly longer to evaluate the
flake's devShell.